Taking Hood out of Shooter
